+++ src/frontend/apps/e2e/__tests__/app-impress/doc-editor.spec.ts
@@ -706,4 +706,69 @@ test.describe('Doc Editor', () => {
       'pink',
     );
   });
+
+  test('it checks interlink feature', async ({ page, browserName }) => {
+    const [randomDoc] = await createDoc(page, 'doc-interlink', browserName, 1);
+
+    await verifyDocName(page, randomDoc);
+
+    const [docChild1] = await createDoc(
+      page,
+      'doc-interlink-child-1',
+      browserName,
+      1,
+      true,
+    );
+
+    await verifyDocName(page, docChild1);
+
+    const [docChild2] = await createDoc(
+      page,
+      'doc-interlink-child-2',
+      browserName,
+      1,
+      true,
+    );
+
+    await verifyDocName(page, docChild2);
+
+    await page.locator('.bn-block-outer').last().fill('/');
+    await page.getByText('Link to a page').first().click();
+
+    await page
+      .locator(
+        "span[data-inline-content-type='interlinkingSearchInline'] input",
+      )
+      .fill('interlink-child-1');
+
+    await page
+      .locator('.quick-search-container')
+      .getByText('interlink-child-1')
+      .click();
+
+    const interlink = page.getByRole('link', {
+      name: 'child-1',
+    });
+
+    await expect(interlink).toBeVisible();
+    await interlink.click();
+
+    await verifyDocName(page, docChild1);
+  });
+
+  test('it checks interlink shortcut @', async ({ page, browserName }) => {
+    const [randomDoc] = await createDoc(page, 'doc-interlink', browserName, 1);
+
+    await verifyDocName(page, randomDoc);
+
+    const editor = page.locator('.bn-block-outer').last();
+    await editor.click();
+    await page.keyboard.press('@');
+
+    await expect(
+      page.locator(
+        "span[data-inline-content-type='interlinkingSearchInline'] input",
+      ),
+    ).toBeVisible();
+  });
 });