+++ src/frontend/apps/impress/src/features/docs/doc-tree/components/DocTree.tsx
@@ -10,8 +10,13 @@ import { css } from 'styled-components';
 
 import { Box, StyledLink } from '@/components';
 import { useCunninghamTheme } from '@/cunningham';
-import { Doc, KEY_SUB_PAGE, useDoc, useDocStore } from '@/docs/doc-management';
-import { SimpleDocItem } from '@/docs/docs-grid';
+import {
+  Doc,
+  KEY_SUB_PAGE,
+  SimpleDocItem,
+  useDoc,
+  useDocStore,
+} from '@/docs/doc-management';
 
 import { useDocTree } from '../api/useDocTree';
 import { useMoveDoc } from '../api/useMove';